* PIECEACHAT.cs: fix insertion
* PIECEVENTE.cs:
* LoginPage.xaml:
* StockMTPage.xaml:
* StockMSPage.xaml:
* StockMEPage.xaml:
* BuyBCEntPage.xaml:
* BuyBCLigPage.xaml:
* BuyFAentPage.xaml:
* BuyFRentPage.xaml:
* SellLignePage.xaml:
* SellFVEntPage.xaml:
* SellFREntPage.xaml:
* SellBREntPage.xaml:
* SellBLEntPage.xaml:
* AffEntetePage.xaml:
* RestServices.cs:
* SellEntetePage.xaml:
* StockLignePage.xaml:
* StockSEentPage.xaml:
* IRestServices.cs:
* BuyBRetEntetPage.xaml:
* BuyBCLigPageModel.cs:
* SellLignePageModel.cs:
* BuyDetailsPageModel.cs:
* SellDetailsPageModel.cs:

+++ PFE/PFE/Models/PIECEACHAT.cs
@@ -33,11 +33,11 @@ namespace PFE.Models
         public int? PCANBPRINT { get; set; }
         public DateTime? PCADATEPRINT { get; set; }
         public int? MODID { get; set; }
-        public int? PCAMNTHT { get; set; }
-        public int? PCAMNTTTC { get; set; }
+        public float? PCAMNTHT { get; set; }
+        public float? PCAMNTTTC { get; set; }
         public int? PCAMNTAREGLER { get; set; }
         public int? PCAMNTACOMPTE { get; set; }
-        public int? PCAMNTTVA { get; set; }
+        public float? PCAMNTTVA { get; set; }
         public int? PCAMNTTPF { get; set; }
         public int? PCAMNTESCOMPTE { get; set; }
         public int? PCAMNTPORT { get; set; }
